Vote part 2
