--- lib/v2/screens/explore_screens/vote/components/proposals/usecases/get_proposals_use_case.dart
@@ -1,9 +1,0 @@
-import 'package:async/async.dart';
-import 'package:seeds/v2/datasource/remote/api/votes_repository.dart';
-import 'package:seeds/v2/screens/explore_screens/vote/interactor/viewmodels/proposal_type_model.dart';
-
-class GetProposalsUseCase {
-  Future<Result> run(ProposalType proposalType) {
-    return VotesRepository().getProposals(proposalType);
-  }
-}
