--- lib/v2/screens/explore_screens/vote/components/proposals/viewmodels/bloc.dart
@@ -1,3 +1,0 @@
-export 'proposals_bloc.dart';
-export 'proposals_event.dart';
-export 'proposals_state.dart';
