--- lib/v2/screens/explore_screens/vote/components/proposals/viewmodels/proposals_event.dart
@@ -1,22 +1,0 @@
-import 'package:equatable/equatable.dart';
-import 'package:meta/meta.dart';
-
-/// --- EVENTS
-@immutable
-abstract class ProposalsEvent extends Equatable {
-  const ProposalsEvent();
-  @override
-  List<Object?> get props => [];
-}
-
-class InitialLoadProposals extends ProposalsEvent {
-  const InitialLoadProposals();
-  @override
-  String toString() => 'InitialLoadProposals';
-}
-
-class LoadProposalsByScroll extends ProposalsEvent {
-  const LoadProposalsByScroll();
-  @override
-  String toString() => 'LoadProposalsByScroll';
-}
