+++ lib/v2/screens/explore_screens/vote_screens/proposals/mappers/proposals_by_scroll_state_mapper.dart
@@ -2,11 +2,11 @@ import 'package:seeds/v2/datasource/remote/model/proposals_model.dart';
 import 'package:seeds/v2/domain-shared/page_state.dart';
 import 'package:seeds/v2/domain-shared/result_to_state_mapper.dart';
 import 'package:seeds/v2/i18n/explore_screens/invite/invite.i18n.dart';
-import 'package:seeds/v2/screens/explore_screens/vote/interactor/viewmodels/proposal_type_model.dart';
-import 'package:seeds/v2/screens/explore_screens/vote/components/proposals/viewmodels/proposals_state.dart';
+import 'package:seeds/v2/screens/explore_screens/vote_screens/proposals/viewmodels/proposals_list_state.dart';
+import 'package:seeds/v2/screens/explore_screens/vote_screens/vote/interactor/viewmodels/proposal_type_model.dart';
 
-class ProposalsStateMapper extends StateMapper {
-  ProposalsState mapResultToState(ProposalsState currentState, Result result) {
+class ProposalsByScrollStateMapper extends StateMapper {
+  ProposalsListState mapResultToState(ProposalsListState currentState, Result result) {
     if (result.isError) {
       return currentState.copyWith(pageState: PageState.failure, errorMessage: "Error loading proposals".i18n);
     } else {
@@ -27,9 +27,11 @@ class ProposalsStateMapper extends StateMapper {
             .toList();
       }
 
-      List<ProposalModel> proposals = currentType.isReverse ? List<ProposalModel>.from(filtered.reversed) : filtered;
+      List<ProposalModel> reversed = currentType.isReverse ? List<ProposalModel>.from(filtered.reversed) : filtered;
+      // Add new proposals to current proposals
+      List<ProposalModel> proposals = currentState.proposals + reversed;
 
-      return currentState.copyWith(pageState: PageState.success, proposals: proposals);
+      return currentState.copyWith(proposals: proposals, hasReachedMax: reversed.isEmpty);
     }
   }
 }
