+++ lib/v2/screens/explore_screens/vote_screens/proposals/viewmodels/bloc.dart
@@ -1,0 +1,3 @@
+export 'proposals_list_bloc.dart';
+export 'proposals_list_event.dart';
+export 'proposals_list_state.dart';
