+++ lib/v2/screens/explore_screens/vote_screens/proposals/viewmodels/proposals_list_event.dart
@@ -1,0 +1,19 @@
+import 'package:meta/meta.dart';
+
+/// --- EVENTS
+@immutable
+abstract class ProposalsListEvent {
+  const ProposalsListEvent();
+}
+
+class InitialLoadProposals extends ProposalsListEvent {
+  const InitialLoadProposals();
+  @override
+  String toString() => 'InitialLoadProposals';
+}
+
+class LoadProposalsByScroll extends ProposalsListEvent {
+  const LoadProposalsByScroll();
+  @override
+  String toString() => 'LoadProposalsByScroll';
+}
