+++ lib/v2/screens/explore_screens/vote_screens/proposals/viewmodels/proposals_list_state.dart
@@ -1,17 +1,17 @@
 import 'package:equatable/equatable.dart';
 import 'package:seeds/v2/datasource/remote/model/proposals_model.dart';
 import 'package:seeds/v2/domain-shared/page_state.dart';
-import 'package:seeds/v2/screens/explore_screens/vote/interactor/viewmodels/proposal_type_model.dart';
+import 'package:seeds/v2/screens/explore_screens/vote_screens/vote/interactor/viewmodels/proposal_type_model.dart';
 
 /// --- STATE
-class ProposalsState extends Equatable {
+class ProposalsListState extends Equatable {
   final PageState pageState;
   final ProposalType currentType;
   final List<ProposalModel> proposals;
   final bool hasReachedMax;
   final String? errorMessage;
 
-  const ProposalsState({
+  const ProposalsListState({
     required this.pageState,
     required this.currentType,
     required this.proposals,
@@ -28,14 +28,14 @@ class ProposalsState extends Equatable {
         errorMessage,
       ];
 
-  ProposalsState copyWith({
+  ProposalsListState copyWith({
     PageState? pageState,
     ProposalType? currentType,
     List<ProposalModel>? proposals,
     bool? hasReachedMax,
     String? errorMessage,
   }) {
-    return ProposalsState(
+    return ProposalsListState(
       pageState: pageState ?? this.pageState,
       currentType: currentType ?? this.currentType,
       proposals: proposals ?? this.proposals,
@@ -44,8 +44,8 @@ class ProposalsState extends Equatable {
     );
   }
 
-  factory ProposalsState.initial(ProposalType proposalType) {
-    return ProposalsState(
+  factory ProposalsListState.initial(ProposalType proposalType) {
+    return ProposalsListState(
       pageState: PageState.initial,
       currentType: proposalType,
       proposals: const [],
