+++ lib/v2/screens/explore_screens/vote_screens/vote/interactor/viewmodels/proposal_type_model.dart
